+++ Libft/ft_strlen_size_t.cpp
@@ -4,7 +4,7 @@
 
 static inline bool has_zero_size_t(size_t value)
 {
-    const size_t mask01 = ~(size_t)0 / 0xFF;
+    const size_t mask01 = ~static_cast<size_t>(0) / 0xFF;
     const size_t mask80 = mask01 << 7;
     return ((value - mask01) & (~value) & mask80) != 0;
 }
@@ -18,7 +18,7 @@ size_t ft_strlen_size_t(const char *string)
     while (reinterpret_cast<uintptr_t>(ptr) & (sizeof(size_t) - 1))
     {
         if (*ptr == '\0')
-            return (ptr - string);
+            return static_cast<size_t>(ptr - string);
         ++ptr;
     }
 
@@ -30,5 +30,5 @@ size_t ft_strlen_size_t(const char *string)
     while (*ptr)
         ++ptr;
 
-    return (ptr - string);
+    return static_cast<size_t>(ptr - string);
 }
